+++ sync-for-expenses/CodatSyncExpenses/ManageData.cs
@@ -48,7 +48,15 @@ namespace CodatSyncExpenses
         /// List pull operations
         /// 
         /// <remarks>
-        /// Gets the pull operation history (datasets) for a given company.
+        /// The *List pull operations* endpoint returns a list of <a href="https://docs.codat.io/sync-for-expenses-api#/schemas/PullOperation">pull operations</a> made by your client.<br/>
+        /// <br/>
+        /// A <a href="https://docs.codat.io/sync-for-expenses-api#/schemas/PullOperation">pull operation</a> is a request to retrieve a specific data type from an integration.<br/>
+        /// <br/>
+        /// ### Tips and traps<br/>
+        /// <br/>
+        /// - The *List pull operations* endpoint does not support querying the `isCompleted` property. You can filter failed pull operations by querying `status!=Complete&amp;&amp;status!=NotSupported` instead.<br/>
+        /// <br/>
+        /// 
         /// </remarks>
         /// </summary>
         Task<ListPullOperationsResponse> ListPullOperationsAsync(ListPullOperationsRequest? request = null);
@@ -85,10 +93,10 @@ namespace CodatSyncExpenses
     {
         public SDKConfig Config { get; private set; }
         private const string _language = "csharp";
-        private const string _sdkVersion = "3.2.0";
-        private const string _sdkGenVersion = "2.159.2";
+        private const string _sdkVersion = "3.3.0";
+        private const string _sdkGenVersion = "2.169.0";
         private const string _openapiDocVersion = "prealpha";
-        private const string _userAgent = "speakeasy-sdk/csharp 3.2.0 2.159.2 prealpha Codat.Sync.Expenses";
+        private const string _userAgent = "speakeasy-sdk/csharp 3.3.0 2.169.0 prealpha Codat.Sync.Expenses";
         private string _serverUrl = "";
         private ISpeakeasyHttpClient _defaultClient;
         private ISpeakeasyHttpClient _securityClient;